+++ dss/util/auth/authorize.py
@@ -1,22 +1,7 @@
-import logging
-
-from dss import Config
-from dss.error import DSSForbiddenException, DSSException
-from dss.util.authorize import fusillade
-
-logger = logging.getLogger(__name__)
-
-
 class Authorize:
-    def __new__(cls, *args, **kwargs):
-        auth_backend = Config.get_auth_backend()
-        if auth_backend == 'FUSILLADE':
-            return fusillade.Fusillade(*args, **kwargs)
-        elif auth_backend == 'AUTH0':
-            # return class for auth0
-            pass
-        else:
-            raise DSSException(500, 'Error with Security Handler, unable to locate Auth Handler')
+    """ abstract class for authorization classes"""
+    def __init__(self):
+        pass
 
     # name this better
     def security_flow(self, authz_methods, *args, **kwargs):
